fixed a few stupid bugs and also added the home page, also updated the dependencies to the latest versions

+++ server/server.js
@@ -244,8 +244,10 @@ app.delete("/servers/:id", authenticate, (req, res) => {
           res.send("Server deleted successfully");
         }
       });
-      terminals[serverId].kill();
-      delete terminals[serverId];
+      if (terminals[serverId]) {
+        terminals[serverId].kill();
+        delete terminals[serverId];
+      }
     }
   });
 });
